Android GUI: Move list diffing code from FreeSpaceFragment to MountInfoAdapter
